// src/payload/payload-types.ts
export interface Job {
  id: number;
  title?: string | null;
  department?: ('academic' | 'admin' | 'support' | 'specialized') | null;
  location?: string | null;
  type?: ('full-time' | 'part-time' | 'contract') | null;
  requirements?:
    | {
        point?: string | null;
        id?: string | null;
      }[]
    | null;
  active?: boolean | null;
  updatedAt: string;
  createdAt: string;
}
// ...
export interface CareerPage {
  id: number;
  hero?: {
    title?: string | null;
    description?: string | null;
  };
  benefits?:
    | {
        iconName?: string | null;
        title?: string | null;
        description?: string | null;
        id?: string | null;
      }[]
    | null;
  updatedAt?: string | null;
  createdAt?: string | null;
}